+++ packages/pymybase/__init__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+"""
+@author: Elias Wood <owns13927@yahoo.com>
+2015-04-05
+base things...
+"""
+
+# from myloggingbase import MyLoggingBase
+# from mydbbase import MyDbBase
+# from myapibase import MyAPIBase
+# from myjson2csv import MyJSON2CSV
+# from mythread import MyThread
+
+__all__ = ['MyLoggingBase','MyJSON2CSV','MyAPIBase','MyDbBase','MyThread']
+
+__title__ = 'pymybase'
+__author__  = 'Elias Wood <owns13927@yahoo.com>'
+__version__ = '0.9.0'
+__date__    = '2017-03-25'